update to extract table title and caption for preprints
Without changing the functionality for extracting data from tables in PMCs, I added code to also allow extraction of title and caption from tables in preprints. Since tables in preprints contain title and caption as text, whereas the rest of the table is an image, they were identified without rows, so the program did not proceed initially.